% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_biplot.R
\name{pca_biplot}
\alias{pca_biplot}
\title{Create a Biplot for Visualizing Principal Component Analysis (PCA) with Grouping}
\usage{
pca_biplot(
  pca_result,
  obs_labels = NULL,
  var_labels = TRUE,
  scale_obs = 1,
  scale_vars = 1,
  arrow_multiplier = 1.5,
  habillage = NULL,
  ...
)
}
\arguments{
\item{pca_result}{A \code{my_prcomp} object. The result of a PCA analysis performed using the \code{my_prcomp} function. This object contains the principal components, loadings, and other relevant data for plotting.}

\item{obs_labels}{A vector of labels for the observations. These labels will be displayed on the plot near the corresponding points. If \code{NULL}, no labels are shown for the observations.}

\item{var_labels}{A logical value indicating whether the names of the variables (loadings) should be displayed as labels on the plot. Default is \code{TRUE}.}

\item{scale_obs}{A numeric value to scale the observations' points on the plot. Adjusts the size of the points representing the observations. Default is \code{1}.}

\item{scale_vars}{A numeric value to scale the variables' arrows on the plot. Adjusts the length of the arrows representing the original variables. Default is \code{1}.}

\item{arrow_multiplier}{A numeric value that controls the length of the arrows representing the variables (loadings). Increasing this value makes the arrows longer. Default is \code{1.5}.}

\item{habillage}{A factor or vector used for grouping the observations. The observations will be color-coded and represented with different shapes based on the levels of this factor. If \code{NULL}, no grouping is applied.}

\item{...}{Additional arguments passed to the underlying plotting functions. This can be used to customize various aspects of the plot, such as colors, themes, and sizes.}
}
\value{
A \code{ggplot} object representing the biplot of the first two principal components. This plot displays observations and variable loadings, with options to color-code observations by groups and label them accordingly.
The plot can be further customized using \code{ggplot2} functions.
}
\description{
Create a Biplot for Visualizing Principal Component Analysis (PCA) with Grouping
}
\examples{
# Perform PCA on a sample dataset
data(iris)
pca_result <- my_prcomp(iris[, -5], scale. = TRUE)

# Basic biplot
pca_biplot(pca_result)

# Biplot with observation labels and custom arrow length
pca_biplot(pca_result, obs_labels = rownames(iris), arrow_multiplier = 2)

# Biplot with grouping by species
pca_biplot(pca_result, obs_labels = rownames(iris), arrow_multiplier = 2, habillage = iris$Species)
}
