% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screeplot_variance.R
\name{screeplot_variance}
\alias{screeplot_variance}
\title{Create a Scree Plot of Explained Variance}
\usage{
screeplot_variance(pca_result, add_labels = TRUE, line = TRUE, ...)
}
\arguments{
\item{pca_result}{A my_prcomp object. The result of a PCA analysis performed using the \code{my_prcomp} function. This object contains the principal components and their associated variance.}

\item{add_labels}{A logical value indicating whether to display the proportion of variance explained by each principal component as labels on the bars. Defaults to \code{TRUE}.}

\item{line}{A logical value indicating whether to include a line representing the cumulative proportion of variance explained. Defaults to \code{TRUE}.}

\item{...}{Additional arguments passed to the ggplot2 plotting functions for further customization.}
}
\value{
A ggplot object representing the scree plot of explained variance for the principal components. The plot can be further customized using \code{ggplot2} functions.
}
\description{
This function creates a scree plot to visualize the proportion of variance explained by each principal component in a PCA. It can also add labels to each bar and optionally include a line to show the cumulative proportion of variance.
}
\examples{
# Perform PCA on a sample dataset
data(iris)
pca_result <- my_prcomp(iris[, -5], scale. = TRUE)

# Basic scree plot
screeplot_variance(pca_result)

# Scree plot with labels and without cumulative line
screeplot_variance(pca_result, add_labels = TRUE, line = FALSE)

# Customize the scree plot using additional ggplot2 arguments
screeplot_variance(pca_result) + ggplot2::theme_minimal()
}
