% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_my_prcomp.R
\name{summary.my_prcomp}
\alias{summary.my_prcomp}
\title{Summary Method for my_prcomp Objects}
\usage{
\method{summary}{my_prcomp}(object, ...)
}
\arguments{
\item{object}{A my_prcomp object. The result of a PCA analysis performed using the my_prcomp function.}

\item{...}{Additional arguments (currently not used).}
}
\value{
This function returns a formatted summary of the PCA results, including the standard deviations of the principal components, the proportion of variance explained by each component, and the cumulative proportion of variance explained.
}
\description{
This function provides a summary of the results from a Principal Component Analysis (PCA) performed using the \code{my_prcomp} function. It calculates and displays the standard deviations, proportion of variance, and cumulative proportion of variance explained by each principal component.
}
\examples{
# Perform PCA on a sample dataset
data(iris)
pca_result <- my_prcomp(iris[, -5], scale. = TRUE)

# Get the summary of the PCA result
summary.my_prcomp(pca_result)
}
